requirements.txt
boto3
kafka-python
ffmpeg-python
python-dotenv
ffmpeg